test(range): migrate states
