+++ core/src/components/range/test/states/range.e2e.ts-snapshots/range-disabled-ios-ltr-Mobile-Safari-linux.png
