+++ core/src/components/range/test/states/range.e2e.ts-snapshots/range-disabled-md-ltr-Mobile-Chrome-linux.png
