+++ core/src/components/range/test/states/range.e2e.ts-snapshots/range-enabled-ios-ltr-Mobile-Chrome-linux.png
