+++ core/src/components/range/test/states/range.e2e.ts-snapshots/range-enabled-ios-ltr-Mobile-Firefox-linux.png
