+++ core/src/components/range/test/states/range.e2e.ts-snapshots/range-enabled-md-ltr-Mobile-Firefox-linux.png
